New 3D GameObject pipeline: ManufactureGO
New pipeline isolates generating GameObjects(GO) from main code.  Elements first get created, and those are 'Manufactured' into a WarehouseGO as well as sent to the GPU.  In the future, if performance needs to be improved, ManufactureGO can be upgraded to do additional optimizations like bursting, and the code is all in one place.  The new pipeline also retains a database of all objects we created, and can be modified and updated in the GPU easily.  We needed this capability for scent overlay.

New Scripts: ElementStore, ManufactureGO, WarehouseGO

New ElementStore.asset

Hooked new pipeline into Build3DFromRooms().