Added grey out only map (not fogs) while in Scent mode for visibility
Prefabs moved to new layer 8 ScentFog.  As a backup, this forced in ManufactureGO.cs:ManufactureInstance()

Commented out a potential enhancement sniffFogBoost, as it requires a Shader edit I'm not sure how to do.  See comments in SniffModeVisuals.cs:SetSniffMode()

Bugfix in RoomPackedAlgorithms.cs: ReportCellDifferences() where cell in wrong direction was used to check adjacent door/wall matching.  Also limited checks to the cardinal directions instead of including diagonals which are not relevant.